blog: update post for v4.7.1
While promoting additional platforms for v4.7.1 after the release, the tarballs
on the release server were overwritten and now have different shasums. In order
to remove any ambiguity around the release we have opted to do a semver patch
release, v4.7.2, with no changes.

This includes a number of things

* a cavaet about v4.7.2
* the download link for sunos
* the original shas
* the shas currently found on nodejs.org