Update release posts (#1097)
* blog: Add sunos x64 to v6.9.2

* blog: Add sunos x64 to v4.7.0

* blog: Add sunos x64 to v4.6.2

* blog: update post for v4.7.1

While promoting additional platforms for v4.7.1 after the release, the tarballs
on the release server were overwritten and now have different shasums. In order
to remove any ambiguity around the release we have opted to do a semver patch
release, v4.7.2, with no changes.

This includes a number of things

* a cavaet about v4.7.2
* the download link for sunos
* the original shas
* the shas currently found on nodejs.org

* blog: update v4.7.2 post

the original post did not include the cavaet

* blog: update post for v6.9.3

While promoting additional platforms for v6.9.3 after the release, the tarballs
on the release server were overwritten and now have different shasums. In order
to remove any ambiguity around the release we have opted to do a semver patch
release, v6.9.4, with no changes.

This includes a number of things

* a cavaet about v6.9.4
* the download link for sunos
* the original shas
* the shas currently found on nodejs.org