use nvm default installation directory
I've tried to use the nvm_version_path function but have failed to get
it to work from within the Makefile. I don't think sourceing nvm.sh from
the Makefile is good as I don't think it should be aware of any package
mangers. The idea of using Make install was that it is just using what
is already there and does not have to modify anything.