minor: update branding for macOS binaries (#1536)
While maintaining the trademark and heritage, Apple no longer actively
refers to any of its Mac platforms as Macintosh. Also, in a number of
places, nodejs.org refers to macOS and Macintosh on the same page.
This change updates the site for consistency, preferring macOS in all
locations. While I've updated the localized versions, the only change
is to revise a single proper noun, with one exception:
`locale/ko/download/package-manager.md` appears to spell "Macintosh" in
Korean, and I am unequipped to handle that change.

This commit does not update blog entries, which refer to the Macintosh
installer by name. Blog entries seem fixed and ephemeral, though it
would be trivial to update them all. Going forward, blog entries
should identify the installer for macOS as the "macOS Installer", as
opposed to the "Macintosh Installer".