Add Node.js User Survey 2018 microsite
This site comes with its own HTML and assets and doesn't depend on any of the
regular website's resources. It's not supposed to be translated.

It is added to a website subdirectory for ease of deployment, we might consider
to move it to a separate subdomain and repository next year, when the next
report comes out.

Fixes: #1654
Review: @gtewallace, @alexcontini