Better GitHub issue templates
The [babel](https://github.com/babel/babel/issues/new/choose) project had a
pretty neat idea for GitHub issue templates, so I'd like to adapt them for
the Node.js website.

As we receive issues for API docs (which are not hosted here) and other, it
might be a good idea to guide the user to the right place directly.