remove unnecessary prose
The text above the list of links on the Get Involved page seems to serve
mostly as a way to disguise the fact that the page is a list of links.
It doesn't really say very much. The links incorporated in the text
occur elsewhere on the page already. Remove the text.