Remove stale resources from get-involved (#1965)
How To Node has not been updated in over two years. A lot has happened
in Node.js in that time. Remove it from the list.

NodeUp is awesome, but there was one episode in all of 2018. It's also
questionable that it occurs in a list with 7 other resources that are
all officially blessed. It gives the impression that NodeUp is official.
It is not. Remove NodeUp from the list.