remove several questionable links from get-involved (#1964)
I'd like to see us weed this page more. There should be a much smaller number of links. Only durable, high-quality links should be listed. Otherwise, it reflects badly on the website and the project. No one will maintain the list and users will have a bad experience.

For example: Right now, if you click on the Nicaraqua Node.js Community link, you will be sent to a placeholder screen that tries to sell you car insurance.

<img width="623" alt="screen shot 2018-12-27 at 9 16 46 am" src="https://user-images.githubusercontent.com/718899/50488601-2facd880-09b8-11e9-8770-ffd715e71baa.png">

Meanwhile, the Vietnam Node.js Community link is broken, the Korean link is questionable (seems to link to an out-of-date version of https://nodejs.org/ko/blog/), the Iranian link seems to be to an even *more* out-of-date version of the website, many of the others don't appear to be community links at all but are people's personal blogs which often have not been updated in years, the Australian site has an invalid cert...

Let someone else curate a living list that tries to be comprehensive. (Maybe someone can start one of those "Awesome" lists on GitHub for Node.js community/getting-involved stuff if there isn't one already.) The website is a poor choice of venue for that kind of thing.

This page needs fewer things, not more, IMO.