remove unnecessary prose (#1966)
* remove unnecessary prose

The text above the list of links on the Get Involved page seems to serve
mostly as a way to disguise the fact that the page is a list of links.
It doesn't really say very much. The links incorporated in the text
occur elsewhere on the page already. Remove the text.

* remove unnecessary prose for zh-CN

remove unnecessary prose for zh-CN

* remove unnecessary prose for zh-TH

remove unnecessary prose for zh-TH

* remove unnecessary prose for ca

remove unnecessary prose for ca

* remove unnecessary prose for es

remove unnecessary prose for es

* remove unnecessary prose for fa

remove unnecessary prose for fa

* remove unnecessary prose for pt-br

remove unnecessary prose for pt-br

* remove unnecessary prose for uk

remove unnecessary prose for uk.

* remove unnecessary prose for fr

remove unnecessary prose for fr