Build default language with locale (adds preserveLocale flag) (#1940)
Currently when navigating to a page that does not have a locale, it will default the english one, this breaks the navigation (it doesn't really break it, but everything including links now changes locale to english)
![2018-12-15 01 11 30](https://user-images.githubusercontent.com/11986564/50038916-c2886300-0007-11e9-8755-0ad42a12c67e.gif)
This PR adds a preserveLocale flag that will run a second build of the files not included in the locale that exist in the "en" locale but using the current locale site.json. 
Build time is considerably longer (*en* locale has almost 700 files, compared to others with 20 or less)
The result:
![2018-12-15 01 16 14](https://user-images.githubusercontent.com/11986564/50039015-86ee9880-0009-11e9-82ca-39405727747b.gif)

Open for discussion.