zh-CN: Fix the anchor links
Consider that some special characters (not included in the 26 English
alphabets) are ignored by md file's anchor when generated. We have to
change the anchor link.

Take this as an example：'SmartOS 和 illumos', the '和' doesn't belong
to an English character, so it's ignored and the generated anchor would
be '#smartos-illumos' instead of '#smartos-and-illumos'.

This is the HOT fix for that. What I did now is to use the pure HTML
instead of md's "##" to force the "a href"'s link to keep it the same
as that in English version.