Fix: Fix for the non-English character generation of Anchor Link
I. In the recent days we've fixed the non-English character generation
of anchor link by directly using the pure HTML tags, however a better
way is:

1) Check whether the current title has the comment symbol (<!---->).
2) If yes, just take the characters out of the symbol as the anchor
link's name, with the real text as the title behind it.
3) If not, just do replacements as what it used to be.

The advantage is that you can keep the anchor link the same as what
it is in the English version or define by yourself. What's more: It's
much more flexible when your change the level of titles, compared with
https://github.com/nodejs/nodejs.org/pull/2028.

II. Add '--fix' for local test to automatically fix some simple
formation typo errors.