Fix: Update the img link of 'schedule.png'
We cannot see the 'schedule.png', so according to this submit
(nodejs/Release@2bf2ea3), it seems the img was removed and we've used
'schedule.svg' instead.

The reason why I use the parameter `?sanitize=true` is that we MUST make
sure that the response type should be of img (Content-Type should be
`img/svg+xml` instead of `text/plain`).

For more about svg rendering on GitHub's server, please see these
related posts:

1. isaacs/github#316 (comment)
2.
https://stackoverflow.com/questions/13808020/include-an-svg-hosted-on-github-in-markdown
(See 'Linking to RAW files (Does not work)').

Fix for: #2055