update: Remove node-events page (#2301)
The page is undermaintained and not very visible anyway. It currently
contains one past event, two upcoming events, and one event that is not
actually happening in 2019. It does not containt the Foundation's own
December conference. We could update it now, but it will simply
immediately go back to being neglected. This is not great content for
the website anyway, in my opinion. If we can get the Get Involved
section slimmed down a fair amount, it might make sense to include a
link or two to external well-maintained sources of information, whether
it be a community-maintained page somewhere or a search on Lanyrd or
similar service. But for now, let's remove this largely unused and
unmaintained page.