blm: convert to banner
We originally discussed keeping the BLM landing page up for a week.
This is a suggested implementation for what we can replace it with.
It makes a new landing page /en/black-lives-matter/ that preserves
the entire original BLM landing site and adds a link to it using our
Security Release notification banner.