Added the minimum criteria for beginners.
As for starting node.js, many gets confused what exactly knowledge and skills required to start this. So if on the first page it is written that a a basic JavaScript and knowledge of callback functions and modules methods is enough to start with node.js, it would give more clearance to the newbies and first time visitor to the websites.