Correct example code (#3408)
The code as written did not work.  I was looking for a terse script to display some local html content, so i googled "node server static".  Although this article is only the 5th result, it should be suitable for copy-paste turnkey.

The article could use some other modernization as well, but I'll leave that out to avoid muddying the main point.