feat: make previous versions of Nodejs pagable
Since we've now got a large table with latest versions of Nodejs, we
don't need previous links but table instead.

What's more, in order to make people download the version quickly, the
new feature is to offer you a quick but simple search for Nodejs's
version, as well as npm version.