remove the list-of-meetups page
This type of content is a poor fit for the website. It has many broken
links, there is no vetting of the quality or sponsorship of the events,
it is impossible to keep up to date, the non-English pages are even more
outdated than the English page, and it is of limited utility.