chore: replace eslint-config-semistandard with eslint-config-standard
eslint-config-semistandard hasn't been published in a year as of this
writing and does not install with the version of eslint that we use
without --force or --legacy-peer-deps.

Replace it with the up-to-date eslint-config-standard, and change the
semi rule in .eslintrc to have the same impact. This also disables
two (new?) rules that now flag errors in our markdown files. These
can be investigated later or not. It also makes a small change to
release-post.js so it complies with a rule about using object
shorthand when possible.

Fixes: https://github.com/nodejs/nodejs.org/issues/4585