fix：Wrong URL to a missing translation page
There're some files that aren't translated into the locale,
which causes the "404 Page Not Found" Error when clicking "Edit on
GitHub".
Now the current solution is: If the file doesn't exist, it will create
a new blank file, with a link of translation guide there to guide what
to do next. The user can either continue translating or abandoning it by
clicking "Cancel" or "Submit".