docs: Fix arabic es6 page layout to rtl
The article direction is set to "auto" , which detect the direction based on the first content. The first content of the page was ES6 which is in latin letters, the page direction is set to "ltr" automatically. I added the word "نسخة" which is the arabic equivalent of "version". It keeps the meaning the same but fixes the layout.