fix: AnchorLinks don't work properly (#5100)
This fixture mainly focuses on two bugs:

1. For Heading <h1>, it doesn't render as what we see for "h2" ,
   "h3"..,and the "id" is always null. So the solution is to render the
   the children (here it must be a string) and return it in default.

2. For very long English titles or non-English titles in md files,
the generated anchor name in default will be very hard to read/write in
md files if we want to refer. What we used to do is to give it an alian
name in short, Take Chinese generated HTML files will be a good example
to explain.