fix: download links cannot be refreshed through js
Problem:

Reason: As far as I see, when using 'querySelectorAll' will store the
whole page and return you the result directly, so the page should be
fully-loaded. Sometimes there's problems when using
"<Script...beforeInactive>" and I change it to "defer async" so as to
speed up the loading of js files as well as to make sure the whole page
is loaded, we can change the link instead of partially loading.