fix: update the translation label for the home page
For the home page of nodejs.org, it seems the download info should be
"labels['download-for']" as an i18n message instead of
"labels['download']", because it fetches the info from "data-dl-local",
whose meaning is the same as "labels['download-for']".

Thus in this md file, the "download" i18n is useless, I've also removed
it out of here.