fix: change to use useDownloadLink hook
The missing filename is due to race condition.
In nodejs.org, the link was rendered in Server Side.
So when that function is invoked, the query selector
can detect it and append the filename.

But this is not the case since we are fetching the
Node releases data in Client Side now. The query
selector will be executed first before Node releases
data is fetched.