Fix: use MDX format to render tables
Changed file extensions to enable processing in the MDX format with the
remark-gfm plugin. This is required to render the HTML tables.
Adapted the Markdown comments and HTML attributes to the MDX format.
The tbody and thead HTML elements were added explicitly to make the HTML
in the server match the one in the client.
No documentation text was changed.