chore: enable MDX type checking
This enables type checking of MDX files. The global `MDXProvidedComponents` type
declares which components are available inside MDX files.

The `**/*.mdx` in `tsconfig.json` makes the MDX files part of the TypeScript
program. This makes TypeScript aware that the MDX files are part of the same
project as TypeScript files and thus share the same context.

The `mdx.checkMdx` enables strict type checking in MDX files. This means type
errors turn into red squiggly lines instead of just providing autocompletion
etc.

At the moment this is only used to enhance the editor experience. The TypeScript
CLI does not check the MDX files (yet).