fix: learn page breadcrumbs where label !== path slug (#6729)
* fix learn page breadcrumbs where label !== path slug

Consider the article "event-loop-timers-and-nexttick" where the label is "The Nodejs Event loop". Breadcrumbs cannot be found in the navigation tree. Ofcourse we can always change the article name to match the label but what about the article "discover-javascript-timers"? Why, there is no problem here! Except there is. The "s" in JavaScript is capitalized. (Shocking, I know. I learned this today). "JavaScript" cannot be written as "java-script" in dash-case. Is it javascript or a script written in Java?

Previously docs should have the same article name and label for the breadcrumbs to work. Now it is not necessary. Unless a deeply nested document structure is adopted, in which case we have to DFS the navigation tree.

Signed-off-by: abizek <abishekilango@protonmail.com>

* test: dashToCamelCase string util

Signed-off-by: abizek <abishekilango@protonmail.com>

* fix: Update nav and i18n keys to match article name

Signed-off-by: abizek <abishekilango@protonmail.com>

---------

Signed-off-by: abizek <abishekilango@protonmail.com>