fix(sidebar): use placeholders in select menu (#7871)
* fix(#7646): add default value if current page is missing in sidebar

sometimes current page can be missing from navigation data if the entry is not present in the navigation.json file. In those cases it shows an empty select box on smaller screens. This PR chooses the first item in the navigation tree if the current page is missing from there

* refactor(#7646): update the select component to have a default placeholder of the window title in site

* refactor(#7646): update the title source to that of frontmatter