fix(blog): Fix homepage link in RSS feed
When following the Node.js Blog at https://nodejs.org/en/blog in a feed reader
like NetNewsWire, the homepage link gets broken. Instead of navigating the
browser to the website, one gets a blank page and an unexpected RSS feed
download in the background.

The `<link>` tag in RSS 2.0 is for the HTML website. Note that the
"self" link from the RSS file back to itself is separate form this
and already auto-generated by the Feed class. [1]

While at it, fix the broken `<description>undefined</defined>` element.
This is often left empty and fine not to pass in `/apps/site/site.json`.
However, the upstream Feed class has a bug where it takes the undefined
and outputs the string "undefined" as the blog's description instead
of a sensible default like empty string.

[1]: https://github.com/jpmonette/feed

Signed-off-by: Timo Tijhof <krinkle@fastmail.com>