fix(learn): Correct unclear example of passing arguments with the task runner
Fix the example used in the 'Passing arguments to the command' section of the
'Run Node.js scripts from the command line' as it is not referencing the right
package.json script.
Also, change the script used in the example ('node app.js --watch') as it can
confuse the reader into thinking it's a way to enable the Node.js watch option
even though it's just passing an '--watch' argument to the app.js script.